OA-388 Show party on Speaker, President & Deputy President profile pages
Back in a4b63b7e and e3f78b23 we stopped saying Speakers and Deputy
Speakers left their party when they "changed party" to enter that
office. Deputy Speakers also had their previous party listed on their
profile but the Speaker didn't because they have to resign from their
party in the UK, which is not the case in Australia.