Disable SSL verification when running mpinfoin script
We were getting "Can't connect to theyvoteforyou.org.au:443 (certificate
verify failed)" errors on the server. I tried a few things to fix this
but gave up and decided to do this instead.

Since we only connect to theyvoteforyou.org.au it's not so bad but it
would be nice to fix it properly one day so I've added a comment to that
effect.

Resolves https://github.com/openaustralia/openaustralia/issues/603