Second sitting week cancelled after Queen's death
Parliament has cancelled its second sitting week for this period. The current media statement is simply that "Federal parliament will not sit as scheduled next week, and will not return for at least a fortnight." - whether that means there will be a replacement sitting week or not, I don't know.